+++ src/components/form/MultifieldGroup/multifieldGroup.stories.js
@@ -55,6 +55,7 @@ FrozenSteps.args = {
   frozen_steps: true,
   items: [
     {
+      placeholders: { name: 'John' },
       values: {
         step_number: '3',
         step_name: 'Cooking chicken',